Code style adjustments; enabled/fixed TestSamplingIterator
git-svn-id: https://svn.apache.org/repos/asf/lucene/mahout/trunk@897440 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mahout/fpm/pfpgrowth/MultiTransactionTreeIterator.java
@@ -51,7 +51,7 @@ public final class MultiTransactionTreeIterator implements Iterator<List<Integer
   public List<Integer> next() {
     List<Integer> returnable = currentPattern.getFirst();
     currentCount++;
-    if (currentCount == currentPattern.getSecond().longValue()) {
+    if (currentCount == currentPattern.getSecond()) {
       if (pIterator.hasNext()) {
         currentPattern = pIterator.next();
         currentCount = 0;
